Improvements on application page. #704

+++ source/src/main/java/org/cerberus/crud/service/impl/CountryEnvironmentParametersService.java
@@ -208,9 +208,9 @@ public class CountryEnvironmentParametersService implements ICountryEnvironmentP
     @Override
     public Answer compareListAndUpdateInsertDeleteElements(String system, String application, List<CountryEnvironmentParameters> newList) {
         Answer ans = new Answer(null);
-        MessageEvent msg = null;
-        msg = new MessageEvent(MessageEventEnum.DATA_OPERATION_OK);
-        msg.setDescription(msg.getDescription().replace("%ITEM%", OBJECT_NAME).replace("%OPERATION%", "UPDATE"));
+
+        MessageEvent msg1 = new MessageEvent(MessageEventEnum.GENERIC_OK);
+        Answer finalAnswer = new Answer(msg1);
 
         List<CountryEnvironmentParameters> oldList = new ArrayList();
         try {
@@ -229,12 +229,16 @@ public class CountryEnvironmentParametersService implements ICountryEnvironmentP
         for (CountryEnvironmentParameters objectDifference : listToUpdateOrInsertToIterate) {
             for (CountryEnvironmentParameters objectInDatabase : oldList) {
                 if (objectDifference.hasSameKey(objectInDatabase)) {
-                    this.update(objectDifference);
+                    ans = this.update(objectDifference);
+                    finalAnswer = AnswerUtil.agregateAnswer(finalAnswer, (Answer) ans);
                     listToUpdateOrInsert.remove(objectDifference);
                 }
             }
         }
-        this.createList(listToUpdateOrInsert);
+        if (!listToUpdateOrInsert.isEmpty()) {
+            ans = this.createList(listToUpdateOrInsert);
+            finalAnswer = AnswerUtil.agregateAnswer(finalAnswer, (Answer) ans);
+        }
 
         /**
          * Delete all objects database Objects that do not exist from newList
@@ -250,8 +254,11 @@ public class CountryEnvironmentParametersService implements ICountryEnvironmentP
                 }
             }
         }
-        ans = this.deleteList(listToDelete);
-        return new Answer(msg);
+        if (!listToDelete.isEmpty()) {
+            ans = this.deleteList(listToDelete);
+            finalAnswer = AnswerUtil.agregateAnswer(finalAnswer, (Answer) ans);
+        }
+        return finalAnswer;
     }
 
     @Override
